Adding mongodb database
