+++ package.json
@@ -19,7 +19,8 @@
   "homepage": "https://github.com/netor27/LibraryWebApp#readme",
   "dependencies": {
     "ejs": "^2.5.7",
-    "express": "^4.16.2"
+    "express": "^4.16.2",
+    "mongodb": "^2.2.33"
   },
   "devDependencies": {
     "gulp": "^3.9.1",
