+++ src/routes/adminRoutes.js
@@ -1,0 +1,72 @@
+var express = require('express');
+var adminRouter = express.Router();
+var mongodb = require('mongodb').MongoClient;
+
+var books = [{
+        title: 'War and Peace',
+        genre: 'Historical Fiction',
+        author: 'Lev Nikolayevich Tolstoy',
+        read: false
+    },
+    {
+        title: 'Les Misérables',
+        genre: 'Historical Fiction',
+        author: 'Victor Hugo',
+        read: false
+    },
+    {
+        title: 'The Time Machine',
+        genre: 'Science Fiction',
+        author: 'H. G. Wells',
+        read: false
+    },
+    {
+        title: 'A Journey into the Center of the Earth',
+        genre: 'Science Fiction',
+        author: 'Jules Verne',
+        read: false
+    },
+    {
+        title: 'The Dark World',
+        genre: 'Fantasy',
+        author: 'Henry Kuttner',
+        read: false
+    },
+    {
+        title: 'The Wind in the Willows',
+        genre: 'Fantasy',
+        author: 'Kenneth Grahame',
+        read: false
+    },
+    {
+        title: 'Life On The Mississippi',
+        genre: 'History',
+        author: 'Mark Twain',
+        read: false
+    },
+    {
+        title: 'Childhood',
+        genre: 'Biography',
+        author: 'Lev Nikolayevich Tolstoy',
+        read: false
+    }
+];
+
+var router = function(nav) {
+
+    adminRouter.route('/addBooks')
+        .get(function(req, res) {
+            var url = 'mongodb://dbuser:M0nG0Us3rP4ssw0rd@ds119345.mlab.com:19345/library-db';
+            mongodb.connect(url, function(err, db) {
+                var collection = db.collection('books');
+                collection.insertMany(books,
+                    function(err, results) {
+                        res.send(results);
+                        db.close();
+                    });
+            });
+        });
+    return adminRouter;
+};
+
+module.exports = router;
